Moves tutorials-related service files into a subfolder in the server folder for the home plugin
